third_party/WebKit/LayoutTests/http/tests/devtools/console/console-copy-treeoutline-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console copies tree outline messages properly.


Running: testSelectAll
Selected text: console-copy-treeoutline.js:16 {foo: 1, bar: "string"}
